CNRLOG and FACTLOG2  , modify power function to long

+++ Main.java
@@ -59,8 +59,8 @@ public class Main
            System.out.print(key+"^"+facs.get(key)+" * ");
        }
        System.out.println();
-       int power= R_power.power(2,9);
-       int Ipower= I_power.power(2,9);
+       long power= R_power.power(2,9);
+       long Ipower= I_power.power(2,9);
        System.out.println("2^9:"+power);
        System.out.println("2^9 Ipower:"+Ipower);
 	   /*balancetree.create(inputnotsort);
